+++ lib/contentful/resource/fields.rb
@@ -1,3 +1,5 @@
+require 'contentful/constants'
+
 module Contentful
   module Resource
     # Include this module into your Resource class to enable it
@@ -7,7 +9,21 @@ module Contentful
     module Fields
       # Returns all fields of the asset
       def fields(wanted_locale = default_locale)
+        wanted_locale = wanted_locale.to_s
         @fields.has_key?(wanted_locale) ? @fields[wanted_locale] : @fields[locale]
+      end
+
+      # Returns all fields of the asset with locales nested by field
+      def fields_with_locales
+        remapped_fields = {}
+        locales.each do |locale|
+          fields(locale).each do |name, value|
+            remapped_fields[name] ||= {}
+            remapped_fields[name][locale.to_sym] = value
+          end
+        end
+
+        remapped_fields
       end
 
       def initialize(object = nil, *)
@@ -24,6 +40,10 @@ module Contentful
       end
 
       private
+
+      def locales
+        @fields.keys
+      end
 
       def extract_fields_from_object!(object)
         initialize_fields_for_localized_resource(object)